Fixed the following bugs 1) fixed grammar error. now expressions are perceived as c-like (previously pascal-like) 2) an error that does not allow you to correctly parse expressions similar to "a; 10;" Added the following features 1) draft version of the poliz performer 2) unary minus

+++ ModelLangInterpreter/Lex.cpp
@@ -1,6 +1,6 @@
 #include "Lex.h"
 
-Lex::Lex(type_of_lex t, unsigned long v, long _num_row_in_file): t_lex(t), v_lex(v),
+Lex::Lex(type_of_lex t, long long v, long _num_row_in_file): t_lex(t), v_lex(v),
                                                                 num_row_in_file(_num_row_in_file){  }
 
 type_of_lex Lex::get_type() const {
